Version 1.0

+++ services/asignaciones/asignaciones.ts
@@ -6,15 +6,25 @@ if (!apiUrl) throw new Error("NEXT_PUBLIC_API_URL is not defined");
 
 
 export const CreateAsignacion = async (
-    data: CreateAsignaciones
+    data: CreateAsignaciones,
+    token: string | undefined
 ) => {
+    if (!token) return;
+
+    const headers = {
+        Authorization: `Bearer ${token}`,
+    };
 
     try{ 
       const result = await axios.post(
         `${apiUrl}/Asignaciones`,
         {
             ...data,
+        },
+        {
+            headers,
         }
+
       );
 
       return result.data;
@@ -60,12 +70,20 @@ export const getAsignacionesByCedula = async (
 }
 
 export const DeleteAsignaciones = async (
-    id: Number
+    id: Number,
+    token: string | undefined
 ) => {
+    if (!token) return;
+
+    const headers = {
+        Authorization: `Bearer ${token}`,
+    };
 
     try{ 
       const result = await axios.delete(
-        `${apiUrl}/Asignaciones/${id}`
+        `${apiUrl}/Asignaciones/${id}`,{
+            headers,
+        }
       );
 
       return result.data;
